Protect getSettings from picking up empty assignments
There are cases where (for whatever reason) PrusaSlicer emits a variable definition at the end with an empty value. For example, in my case it was emitting:

   ; printer_model = 

This line, when processed via `getSettings()`, results in `"", true`. Notice the `true` at the end. It thinks this is a legitimate value when in reality it's just empty.

The problem is that when one tries to "overwrite" this in PrusaSlicer, you can only put custom G-code _before_ the variables dump at the end. Therefore, no matter what you do, as long as getSettings() returns a `true`, the value that we emit gets overwritten by an empty string, and in my case it results in a "invalid G-code"